requirements.txt
crewai
langchain_openai
streamlit
python-dotenv
langchain_community
groq